clean up create pod modal (#1059)
