+++ wondrous-app/components/Common/AndroidSwitch.tsx
@@ -1,0 +1,35 @@
+import { styled, Switch } from '@mui/material';
+
+const AndroidSwitch = styled(Switch)(({ theme }) => ({
+  padding: 8,
+  '& .MuiSwitch-track': {
+    borderRadius: 22 / 2,
+    background: '#3E3E3E',
+
+    '&:before': {
+      content: '""',
+      position: 'absolute',
+      top: '50%',
+      transform: 'translateY(-50%)',
+      width: 16,
+      height: 16,
+      backgroundImage: `url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" height="16" width="16" viewBox="0 0 24 24"><path fill="white" d="M21,7L9,19L3.5,13.5L4.91,12.09L9,16.17L19.59,5.59L21,7Z"/></svg>')`,
+      left: 12,
+      zIndex: 1000,
+      opacity: 1,
+    },
+  },
+  '& .MuiSwitch-thumb': {
+    boxShadow: 'none',
+    width: 16,
+    height: 16,
+    margin: 2,
+    background: 'white',
+  },
+
+  '& .MuiSwitch-colorSecondary.Mui-checked + .MuiSwitch-track': {
+    background: 'linear-gradient(270deg, #CCBBFF -5.62%, #7427FF 45.92%, #00BAFF 103.12%)',
+  },
+}));
+
+export default AndroidSwitch;
