+++ wondrous-app/components/Pod/analytics/index.tsx
@@ -40,7 +40,6 @@ import {
   StyledChip,
 } from 'components/CreateEntity/styles';
 import palette from 'theme/palette';
-import { filterOrgUsers } from 'components/CreateEntity/CreatePodModal';
 import CSVModal from 'components/organization/analytics/CSVModal';
 import { calculateCount, exportContributorTaskCSV, getContributorTaskData } from 'components/organization/analytics';
 import { BOUNTY_TYPE, PRIVATE_TASK_TITLE } from 'utils/constants';
